src/inference/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
efficientnet_pytorch
faiss-cpu
azure-ai-ml
azure-identity
